Move Initials logic into ViewModel
